[contrib][linux] Add huf_decompress_amd64.o target to Makefile
Commit 50fe2aa24ee0 ("Huffman ASM") added a new ASM source file,
but it wasn't added to the kernel Makefile despite that it received
support for Huffman ASM according to the internal definitions. This
leads to undefined references, as huf_decompress.o now calls those
ASM functions.
Add it to the list of sources when building inside the kernel tree.
Kbuild can handle .S files just fine, so none additional rules
needed.

Fixes: 50fe2aa24ee0 ("Huffman ASM")
Signed-off-by: Alexander Lobakin <alobakin@pm.me>